Added back the live scan page
I added back the live scan page as an additional page on the dev page.

Currently, it uses the old model, and it is very unoptimized, so it loses frame rate very quickly.

I also tried to fix some initialization errors, which... most of them seem to be gone now.