Implemented LEGO scan model using open source code
Using an open source Flutter tutorial project, I was able to implement their code in order to allow our model that uses Teachable Machine to work within our app. At the moment, it has essentially been copied straight into the app, which of course is not what we want. But this is a good demo for us to learn from and implement in our own way.