Commented new code and changed UI
I went and commented some of the code for the UI widgets for the photo scanner.

I also changed most of the UI, including the colors, and layout of the image and text, and all of the text.

I'm planning on continuing to look over all of this to comment and understand it further.