Implemented Working Scan History Functionality Prototype
The scan history feature is now working using the new TFLite scan implementation.

This was done using the Provider package rather than using the traditional StatefulWidget or setState() methods.

Scanning a LEGO will now update the lists of the History ListView structure with the new Image, Brick Name, and the Accuracy result in place of the color.

Currently, the UI still looks like butt, and the Accuracy values look a little wack, and failed scan results are still saved (which I will change later), it is currently functional!

Let me know if y'all have any questions on how this works! I'll try to help as best I can.