Added a working Bottom Navigation Bar
There is now a working bottom navigation bar in the app.

The main() function now calls BottomNav() as the home instead of LegoRecogniser(). BottomNav() is a class that creates a widget of a bottom navigation bar, that allows the user to switch between the currently selected page. Tapping an icon on the bottom navigation bar will change, the currentIndex, which will then display a new body widget (a page), based off of the screens final list.

State is also maintained with this, so as far as I can tell, it is fully functional!